Refactor JSON serialization into CycloneDX.Json library
