+++ CycloneDX.Models/License.cs
@@ -14,13 +14,8 @@
 //
 // Copyright (c) Steve Springett. All Rights Reserved.
 
-using System.Text.Json.Serialization;
-
-using CycloneDX.JsonConverters;
-
 namespace CycloneDX.Models
 {
-    [JsonConverter(typeof(LicenseConverter))]
     public class License
     {
         public string Id { get; set; }
